% Generated by build script
\name{nom_reverse}
\alias{nom_reverse}
\title{Reverse geocode — coordinates to address}
\usage{
nom_reverse()
}
\arguments{
\item{latitude}{Latitude}
\item{longitude}{Longitude}
}
\value{tibble: one row with place_id, osm_id, display_name, latitude, longitude, type, importance}
\description{Reverse geocode — coordinates to address}
