% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nom.R
\name{nom_reverse}
\alias{nom_reverse}
\title{Reverse geocode — coordinates to address}
\usage{
nom_reverse(latitude, longitude)
}
\arguments{
\item{latitude}{Latitude}

\item{longitude}{Longitude}
}
\value{
tibble: one row with place_id, osm_id, display_name, latitude, longitude, type, importance
}
\description{
Reverse geocode — coordinates to address
}
